Allow to disable selected index by toggling it again

+++ src/main/java/mchorse/mclib/client/gui/framework/elements/input/GuiTrackpadElement.java
@@ -76,7 +76,7 @@ public class GuiTrackpadElement extends GuiElement implements IFocusedGuiElement
         return this;
     }
 
-    public GuiTrackpadElement increment(int increment)
+    public GuiTrackpadElement increment(float increment)
     {
         this.increment = increment;
